break everything into own file - debug todo-fix blocks attr id
