+++ .eslintrc
@@ -16,7 +16,7 @@
     "consistent-return": "warn",
     "no-unused-vars": "warn",
     "no-func-assign": "warn",
-    // "no-undef": "warn",
-    "no-use-before-define": "off"
+    "no-use-before-define": "off",
+    "func-names": "off"
   }
 }
